+++ pkgs/http/test/html/client_conformance_test.dart
@@ -3,14 +3,16 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @TestOn('browser')
+library;
 
 import 'package:http/browser_client.dart';
 import 'package:http_client_conformance_tests/http_client_conformance_tests.dart';
 import 'package:test/test.dart';
 
 void main() {
-  testAll(() => BrowserClient(),
+  testAll(BrowserClient.new,
       redirectAlwaysAllowed: true,
       canStreamRequestBody: false,
-      canStreamResponseBody: false);
+      canStreamResponseBody: false,
+      canWorkInIsolates: false);
 }