+++ pkgs/http_client_conformance_tests/lib/http_client_conformance_tests.dart
@@ -6,6 +6,7 @@ import 'package:http/http.dart';
 
 import 'src/close_tests.dart';
 import 'src/compressed_response_body_tests.dart';
+import 'src/isolate_test.dart';
 import 'src/multiple_clients_tests.dart';
 import 'src/redirect_tests.dart';
 import 'src/request_body_streamed_tests.dart';
@@ -19,6 +20,7 @@ import 'src/server_errors_test.dart';
 export 'src/close_tests.dart' show testClose;
 export 'src/compressed_response_body_tests.dart'
     show testCompressedResponseBody;
+export 'src/isolate_test.dart' show testIsolate;
 export 'src/multiple_clients_tests.dart' show testMultipleClients;
 export 'src/redirect_tests.dart' show testRedirect;
 export 'src/request_body_streamed_tests.dart' show testRequestBodyStreamed;
@@ -42,13 +44,17 @@ export 'src/server_errors_test.dart' show testServerErrors;
 /// If [redirectAlwaysAllowed] is `true` then tests that require the [Client]
 /// to limit redirects will be skipped.
 ///
+/// If [canWorkInIsolates] is `false` then tests that require that the [Client]
+/// work in Isolates other than the main isolate will be skipped.
+///
 /// The tests are run against a series of HTTP servers that are started by the
 /// tests. If the tests are run in the browser, then the test servers are
 /// started in another process. Otherwise, the test servers are run in-process.
 void testAll(Client Function() clientFactory,
     {bool canStreamRequestBody = true,
     bool canStreamResponseBody = true,
-    bool redirectAlwaysAllowed = false}) {
+    bool redirectAlwaysAllowed = false,
+    bool canWorkInIsolates = true}) {
   testRequestBody(clientFactory());
   testRequestBodyStreamed(clientFactory(),
       canStreamRequestBody: canStreamRequestBody);
@@ -63,4 +69,5 @@ void testAll(Client Function() clientFactory,
   testCompressedResponseBody(clientFactory());
   testMultipleClients(clientFactory);
   testClose(clientFactory);
+  testIsolate(clientFactory, canWorkInIsolates: canWorkInIsolates);
 }